Expose GE Z-Wave in wall fan switch as a fan
Wink recognizes this as a dimmer switch for some reason. But in
reality it is just a fan switch with 3 speeds (translating to
brightness levels of 33%, 66% and 100% in the dimmer switch).

This patch extends WinkFan to specialize the case of GE Z-Wave
in wall fan switch and exposes it as a fan instead of a light
bulb. I've also added a sample json for the tests.

This way home assistant will expose this as a fan without any
changes on the user's part.